pyproject.toml
[project]
name = "nam-risk-scoring"
version = "0.1.0"
description = "Neural Additive Models for Interpretable Credit Risk with Uncertainty Quantification"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Deep learning
    "torch>=2.1.0",
    # Gradient boosting baseline
    "xgboost>=2.0.0",
    "shap>=0.43.0",
    "numba>=0.59.0",
    # Conformal prediction
    "mapie>=0.8.0",
    # Data and ML utilities
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    # Configuration
    "pyyaml>=6.0",
    # Experiment tracking
    "mlflow>=2.8.0",
    # Calibration
    "netcal>=1.3.0",
    # Jupyter
    "jupyterlab>=4.0.0",
    "ipywidgets>=8.0.0",
    # Hyperparameter optimization
    "optuna>=3.4.0",
    # Testing
    "pytest>=7.4.0",
    # Misc
    "tqdm>=4.66.0",
    "ucimlrepo>=0.0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
